requirements.txt
selenium==4.22.0
python-docx==1.1.2
schedule==1.2.2
aspose-words==24.7.0
